split out channel connectivity test
